move scrap-specific functons to own file
