+++ Content.Server/_MORBIT/Scraps/EntitySystems/ScrapSystem.cs
@@ -40,23 +40,6 @@ public sealed partial class ScrapSystem : SharedScrapSystem
             UpdateMotifs();
     }
 
-    // If you are a scrap, you can always see scraps. I guess
-    private void OnScrapGetVisMask(Entity<ScrapComponent> ent, ref GetVisMaskEvent args)
-    {
-        if (ent.Comp.LifeStage <= ComponentLifeStage.Running)
-            args.VisibilityMask |= (int)VisibilityFlags.Scrap;
-    }
-
-    private void OnScrapStartup(Entity<ScrapComponent> ent, ref ComponentStartup args)
-    {
-        AddToScrapLayer(ent);
-    }
-
-    private void OnScrapShutdown(Entity<ScrapComponent> ent, ref ComponentShutdown args)
-    {
-        RemoveFromScrapLayer(ent);
-    }
-
     private void AddToScrapLayer(EntityUid uid)
     {
         if (Terminating(uid))
